Add files via upload
First version of the "legs.gd" script; used for adding basic 3d movement to CharacterBody3D's.

This is specifically designed for use as follows: attach this as the script of a Node3D, which is directly a child of your object. It hasn't been fully tested outside of basic character-play.