Create dialogue_manager.gd
This is a helper code-snippet, made to store lines of dialogue. The basic format is characterName::message. For instance, if I type in, "Mr. Rogers::Good morning, neighbor." from the editor, then the function "_parseDialogue()" will return "Mr. Rogers" or "Good morning, neighbor.", depending on the given parameters, as a string.