Create ui_menu.gd
This is to work with the dialogue_manager code. This script creates dialogue boxes and fills them with text based on the dialogue_managers.

It is... still a work in progress, I won't lie. But then, it's also my first time sitting down and trying to make a dialogue system.