Create health_heart
Control-type node, made to be a child of the player character, though it never directly accesses the character. This script is both the health And the visual for health, and includes programmatic generation of a health-bar. Currently, only "hearts" are implemented; next would be a health-bar type thing, as I think it would be easier to code.